// planning-microfrontend/src/app/shared/models/event.ts
export class Event {
    eventId: string;
    eventLabel: string;
    eventDate: string;
    eventTime:string;
    eventDesc: string;

    // constructor(eventLabel:string,eventDate:string,eventDesc:string){
    //     this.eventLabel=eventLabel;
    //     this.eventDate=eventDate;
    //     this.eventDesc=eventDesc;
    // }

    

}